+++ lib/spid/generate_authn_request.rb
@@ -5,16 +5,14 @@ require "onelogin/ruby-saml/settings"
 
 module Spid
   class GenerateAuthnRequest # :nodoc:
-    attr_reader :authn_request_attributes
+    attr_reader :authn_request_attributes,
+                :service_provider_configuration
 
     # rubocop:disable Metrics/MethodLength
     # rubocop:disable Metrics/ParameterLists
     def initialize(
           idp_sso_target_url:,
-          assertion_consumer_service_url:,
-          private_key_filepath:,
-          certificate_filepath:,
-          issuer:,
+          service_provider_configuration:,
           authn_context: Spid::L1,
           authn_context_comparison: Spid::EXACT_COMPARISON,
           digest_method: Spid::SHA256,
@@ -33,21 +31,22 @@ module Spid
               " use one of #{COMPARISON_METHODS.join(', ')}"
       end
 
+      @service_provider_configuration = service_provider_configuration
       @authn_request_attributes = {
         idp_sso_target_url: idp_sso_target_url,
         assertion_consumer_service_url: assertion_consumer_service_url,
         protocol_binding: "urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST",
         issuer: issuer,
-        private_key: File.read(private_key_filepath),
-        certificate: File.read(certificate_filepath),
+        private_key: private_key_content,
+        certificate: certificate_content,
         name_identifier_format: name_identifier_format,
         authn_context: authn_context,
         authn_context_comparison: authn_context_comparison,
         security: {
           authn_requests_signed: true,
           embed_sign: true,
-          digest_method:             digest_method,
-          signature_method:          signature_method
+          digest_method: digest_method,
+          signature_method: signature_method
         }
       }
 
@@ -59,6 +58,22 @@ module Spid
 
     def to_saml
       authn_request.create(saml_settings)
+    end
+
+    def assertion_consumer_service_url
+      service_provider_configuration.sso_url
+    end
+
+    def issuer
+      service_provider_configuration.host
+    end
+
+    def private_key_content
+      service_provider_configuration.private_key
+    end
+
+    def certificate_content
+      service_provider_configuration.certificate
     end
 
     private